requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
pandas>=2.2.0
requests>=2.32.0
streamlit>=1.32.0
pyngrok>=7.1.0
pytest>=8.2.0
MetaTrader5>=5.0.0
